coproc/tests: Moving simple tests out of router_tests.cc
- Tests that didn't use the more complex benchmark move into a new file called
'pacemaker_tests.cc' and they also use the coproc_test_fixture instead of the
router_test_fixture
